changed long types to int.

+++ Base_Framework/Domain/Core/Contracts/IRepository.cs
@@ -7,7 +7,7 @@ namespace Base_Framework.Domain.Core.Contracts
     {
         public bool IsExist(Expression<Func<T, bool>> expression);
         public void Save();
-        public void Remove(long id);
-        public void Restore(long id);
+        public void Remove(int id);
+        public void Restore(int id);
     }
 }
