+++ Help.Domain.Core/HelpServiceAgg/DTOs/Comment/SearchCommentDTO.cs
@@ -4,7 +4,7 @@ namespace Help.Domain.Core.HelpServiceAgg.DTOs.Comment
 {
     public class SearchCommentDTO
     {
-        public long HelpRequestId { get; set; }
+        public int HelpRequestId { get; set; }
         public string Message { get; set; }
     }
 }
