+++ Help.Domain.Core/HelpServiceAgg/DTOs/Skill/SkillDTO.cs
@@ -6,12 +6,12 @@ namespace Help.Domain.Core.HelpServiceAgg.DTOs.Skill
 {
     public class SkillDTO
     {
-        public long Id { get; set; }
+        public int Id { get; set; }
         public string CreationDate { get; set; }
         public string Title { get; set; }
         public string Description { get; set; }
         public Int16 Level { get; set; }
         public IdTitleHelpServiceDTO HelpService { get; set; }
-        public long CustomerId { get; set; } 
+        public int CustomerId { get; set; } 
     }
 }
